@@ -12,19 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build !linux
-// +build !linux
+//go:build !linux && !windows
 
 package notify
 
-func notifyReadiness() error {
-	return nil
-}
-
-func notifyReloading() error {
-	return nil
-}
-
-func notifyStopping() error {
-	return nil
-}
+func notifyReadiness() error { return nil }
+func notifyReloading() error { return nil }
+func notifyStopping() error  { return nil }
